Show how to use developer mode
Took a bit of debugging to figure out you must pass the `developerMode` param so that it uses the correct endpoint for a test account. I was also interested in using Authorize.net, here is the sample code I came up with that made it work. Maybe you don't merge this PR but could at least put the bit about the developerMode somewhere in the readme?