Annotations for optional methods on GatewayInterface
This should help IDE's with code completion and suggestions on the gateway methods and return values (which doesn't work very well at the moment).

I understand that these methods are optional - so this may not be the right way to address this issue. My personal opinion is that documenting the method as optional in the annotation is better than not documenting it all.

There are also lots of methods in the AbstractGateway that I think should be included in the GatewayInterface (supportsAuthorize, supportsCompleteAuthorize, supportsCapture, etc...). I wanted to know peoples thoughts on this before I spent the time on a PR.