Document how to implement test mode
Given that implementers of Omnipay are looking to have one piece of code work with multiple gateways testMode provides some challenges. 

I'm not sure the best way to deal with this. I would argue that Authorize.net should implement setTestMode to mean 'use sandbox account' since that is the normal meaning for other gateways, but that would not be backward compatible. 

At minimum I thought I could document how people should implement this to achieve the desired result